move test to release folder (#492)
Category: feature
JIRA issue: MIC-5866

Move test_column_dtypes to release suite framework.

Testing
Ran tests on 5 datasets (acs, cps, wic, ssa and census)
